seeker/src/Metam/requirements.txt
scikit-learn
pandas
TPOT
auto-sklearn
scikit-learn 
causal-learn
elasticsearch
networkx
bitarray
psutil
tqdm
